@@ -1,7 +1,6 @@
 package com.live.adpter;
 
 import android.content.Intent;
-import android.support.annotation.Nullable;
 import android.view.View;
 import android.widget.TextView;
 
@@ -13,6 +12,8 @@ import com.live.entity.Anchor;
 import com.live.ui.juhe.BDLiveActivity;
 
 import java.util.List;
+
+import androidx.annotation.Nullable;
 
 public class AnchorAdapter extends BaseQuickAdapter<Anchor.ZhuboBean, BaseViewHolder> {
 
